requirements.txt
fastapi[standard]
uvicorn==0.30.1
pydantic==2.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pysentimiento==0.7.3
google-genai==0.3.0